+++ uint96/uint96.go
@@ -227,7 +227,7 @@ func (u *Uint96) Copy() Uint96 {
 	}
 }
 
-func (u *Uint96) SetBit(i uint, b uint) {
+func (u *Uint96) SetBit(i uint8, b uint8) {
 
 	if b == 1 {
 		if i >= 64 {
@@ -254,16 +254,16 @@ func (u *Uint96) SetBit(i uint, b uint) {
 	}
 }
 
-func (u *Uint96) GetBit(i uint) uint {
+func (u *Uint96) GetBit(i uint8) uint8 {
 	if i < 32 {
-		return (uint)(u.Lo>>i) & 1
+		return (uint8)(u.Lo>>i) & 1
 	}
 
 	if i >= 32 && i < 64 {
-		return (uint)(u.Mid>>(i-32)) & 1
+		return (uint8)(u.Mid>>(i-32)) & 1
 	}
 
-	return (uint)(u.Hi>>(i-64)) & 1
+	return (uint8)(u.Hi>>(i-64)) & 1
 }
 
 // TrailingZeros returns the number of trailing zero bits in u; the result is
